add Paginate158 method
PaginateToken does not strictly adhere to aip.dev/158, and allows for any type of token. Additionally, the signal that the end of collection has been reached is that the next page function returns an emtpy list; however aip.dev/158 requires that the token being an empty string is the only way to signal this.